+++ tests/integration/main.rs
@@ -446,7 +446,7 @@ async fn test_p2idr_transfer_consumed_by_sender() {
         matches!(
             err,
             ClientError::TransactionExecutorError(
-                TransactionExecutorError::ExecuteTransactionProgramFailed(_)
+                TransactionExecutorError::TransactionProgramExecutionFailed(_)
             )
         )
     }));
@@ -1356,9 +1356,7 @@ async fn test_custom_transaction_prover() {
             &self,
             _tx_witness: TransactionWitness,
         ) -> Result<ProvenTransaction, TransactionProverError> {
-            return Err(TransactionProverError::InternalError(
-                "This prover always fails".to_string(),
-            ));
+            return Err(TransactionProverError::other("This prover always fails"));
         }
     }
 
@@ -1392,7 +1390,10 @@ async fn test_custom_transaction_prover() {
 
     assert!(matches!(
         result,
-        Err(ClientError::TransactionProvingError(TransactionProverError::InternalError(_)))
+        Err(ClientError::TransactionProvingError(TransactionProverError::Other {
+            error_msg: _,
+            source: _
+        }))
     ));
 }